Undo epytext changes and add section id slugification in node2stan.py
